Move getMostTopStartPosition to the helpers,  update tableFocusExit description, add afterDialogFocus tests
